backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
pydantic==2.10.6
python-dotenv==1.0.1
